requirements.txt
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
defusedxml==0.7.1
distlib==0.3.1
Django==3.1.2
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
env==0.1.0
filelock==3.0.12
httpie==2.4.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
nodeenv==1.5.0
oauthlib==3.1.0
pipenv==2020.11.15
psycopg2==2.8.6
pycparser==2.20
Pygments==2.7.4
PyJWT==2.0.1
PySocks==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
virtualenv==20.4.0
virtualenv-clone==0.5.4
